Backend/requirements.txt
Django>=4.2,<4.3
django-oscar>=3.2,<4.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<26.0
django-environ>=0.11,<1.0
django-cors-headers>=4.4,<5.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
opensearch-py>=2.5,<3.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.0,<12.0
sorl-thumbnail>=12.10,<13.0
torch>=2.2,<3.0
transformers>=4.40,<5.0
